@@ -1,11 +1,9 @@
 # Copyright (c) 2017 Vault12, Inc.
 # MIT License https://opensource.org/licenses/MIT
 
-require 'utils'
-
-class ZAXCommand
+class ZaxCommand
   include Utils
-  include NonceHelper
+  include Helpers::NonceHelper
 
   def initialize(hpk,mailbox)
     @hpk = hpk
