@@ -1,8 +1,8 @@
 # Copyright (c) 2015 Vault12, Inc.
 # MIT License https://opensource.org/licenses/MIT
-#
+
 # Load the Rails application.
-require File.expand_path('../application', __FILE__)
+require_relative "application"
 
 Encoding.default_external = Encoding::UTF_8
 # Initialize the Rails application.
