@@ -13,11 +13,14 @@ https://groups.google.com/forum/#!topic/redis-db/T7JzYqYEAqk
 require 'key_params'
 include KeyParams
 
-Redis.current = Redis.new host: 'localhost', port: 6379
+$redis = Redis.new(
+  url: ENV.fetch("REDIS_URL") { "redis://localhost:6379/0" },
+  ssl_params: { verify_mode: OpenSSL::SSL::VERIFY_NONE }
+)
 
 # Used by file manager to keep watch on files to be deleted later
-Redis.current.persist ZAX_GLOBAL_FILES
+$redis.persist ZAX_GLOBAL_FILES
 
 # Used by session difficulty throttling
-Redis.current.set ZAX_ORIGINAL_DIFF, Rails.configuration.x.relay.difficulty
-Redis.current.persist ZAX_ORIGINAL_DIFF
+$redis.set ZAX_ORIGINAL_DIFF, Rails.configuration.x.relay.difficulty
+$redis.persist ZAX_ORIGINAL_DIFF
