@@ -19,6 +19,6 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 0) do
+ActiveRecord::Schema[6.1].define(version: 0) do
 
 end
