@@ -1,11 +1,13 @@
 # Copyright (c) 2015 Vault12, Inc.
 # MIT License https://opensource.org/licenses/MIT
 require 'errors/zax_error'
+
 module Errors
-  class SevereKeyError < ZAXError
+  class RelayTokenError < ZaxError
     def http_fail
       super
-      severe_error 'NaCl key generation error'
+      warn "#{INFO_NEG} bad relay token\n"\
+      "#{@data[:msg]} where relay token = #{dump @data[:relay_token]}"
     end
   end
 end
