@@ -3,10 +3,7 @@
 require 'errors/zax_error'
 
 module Errors
-  class HPKError < ZAXError
-    def http_fail
-      super
-      warn "#{INFO_NEG} hpk error: #{@data[:msg]} #{dump @data[:hpk]}"
-    end
+  class ReportError < ZaxError
+
   end
 end
