@@ -1,13 +1,11 @@
 # Copyright (c) 2015 Vault12, Inc.
 # MIT License https://opensource.org/licenses/MIT
 require 'errors/zax_error'
-
 module Errors
-  class RelayTokenError < ZAXError
+  class ServerKeyError < ZaxError
     def http_fail
       super
-      warn "#{INFO_NEG} bad relay token\n"\
-      "#{@data[:msg]} where relay token = #{dump @data[:relay_token]}"
+      severe_error 'NaCl key generation error'
     end
   end
 end
