@@ -2,7 +2,7 @@
 # MIT License https://opensource.org/licenses/MIT
 require 'errors/zax_error'
 module Errors
-  class SevereRandomError < ZAXError
+  class SeverRandomError < ZaxError
     def http_fail
       @response_code = :internal_server_error
       super
