@@ -1,17 +1,11 @@
 # Copyright (c) 2015 Vault12, Inc.
 # MIT License https://opensource.org/licenses/MIT
-require 'errors/all'
-require 'helpers/request_body_helper'
-require 'helpers/client_token_helper'
-require 'helpers/hpk_helper'
-require 'helpers/nonce_helper'
-require 'helpers/transaction_helper'
 
 module ResponseHelper
   include Errors
-  include RequestBodyHelper
-  include ClientTokenHelper
-  include HPKHelper
-  include NonceHelper
-  include TransactionHelper
+  include Helpers::RequestBodyHelper
+  include Helpers::ClientTokenHelper
+  include Helpers::HpkHelper
+  include Helpers::NonceHelper
+  include Helpers::TransactionHelper
 end
